# api/my_private_finances/api/routes/categories.py
from __future__ import annotations

from typing import Annotated

from fastapi import APIRouter, Body, HTTPException
from sqlmodel import select

from my_private_finances.deps import SessionDep
from my_private_finances.models import Category, Transaction, TransactionSplit
from my_private_finances.schemas import CategoryCreate, CategoryRead, CategoryUpdate

router = APIRouter(prefix="/categories", tags=["categories"])


@router.post("", response_model=CategoryRead, status_code=201)
async def create_category(
    category: Annotated[CategoryCreate, Body()], session: SessionDep
) -> Category:
    if category.parent_id is not None:
        parent = await session.get(Category, category.parent_id)
        if parent is None:
            raise HTTPException(status_code=422, detail="parent_id does not exist")

    db_obj = Category(
        name=category.name,
        parent_id=category.parent_id,
        cost_type=category.cost_type,
    )
    session.add(db_obj)
    await session.commit()
    await session.refresh(db_obj)
    return db_obj


@router.get("", response_model=list[CategoryRead])
async def list_categories(session: SessionDep) -> list[Category]:
    res = await session.execute(select(Category).order_by(Category.name))  # type: ignore[arg-type]
    return list(res.scalars().all())


@router.patch("/{category_id}", response_model=CategoryRead)
async def update_category(
    category_id: int,
    payload: Annotated[CategoryUpdate, Body()],
    session: SessionDep,
) -> Category:
    db_obj = await session.get(Category, category_id)
    if db_obj is None:
        raise HTTPException(status_code=404, detail="Category not found")

    if payload.name is not None:
        db_obj.name = payload.name

    if "cost_type" in payload.model_fields_set:
        db_obj.cost_type = payload.cost_type

    if payload.parent_id is not None:
        if payload.parent_id == category_id:
            raise HTTPException(
                status_code=422, detail="Category cannot be its own parent"
            )
        parent = await session.get(Category, payload.parent_id)
        if parent is None:
            raise HTTPException(status_code=422, detail="parent_id does not exist")
        db_obj.parent_id = payload.parent_id

    await session.commit()
    await session.refresh(db_obj)
    return db_obj


@router.delete("/{category_id}", status_code=204)
async def delete_category(category_id: int, session: SessionDep) -> None:
    db_obj = await session.get(Category, category_id)
    if db_obj is None:
        raise HTTPException(status_code=404, detail="Category not found")

    # Check if any transactions reference this category
    # Transactions are the one relationship we refuse to touch implicitly — the
    # user must re-categorise or accept losing the categorisation first. Every
    # other reference is handled by the FK's ON DELETE (#117): budgets and rules
    # CASCADE, sub-categories and recurring patterns SET NULL.
    result = await session.execute(
        select(Transaction.id).where(Transaction.category_id == category_id).limit(1)
    )
    if result.scalar_one_or_none() is not None:
        raise HTTPException(
            status_code=409,
            detail="Category is in use by transactions and cannot be deleted",
        )

    split_result = await session.execute(
        select(TransactionSplit.id)
        .where(TransactionSplit.category_id == category_id)
        .limit(1)
    )
    if split_result.scalar_one_or_none() is not None:
        raise HTTPException(
            status_code=409,
            detail="Category is in use by a transaction split and cannot be deleted",
        )

    await session.delete(db_obj)
    await session.commit()